% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_sinasc.R
\docType{data}
\name{dados_sinasc}
\alias{dados_sinasc}
\title{Dados do SINASC (Sistema de Informações sobre Nascidos Vivos)}
\format{
forneça uma visão geral do conjunto de dados. Para `data.frames`,
 você deve incluir uma lista de definições que descreva cada variável.
 Geralmente é uma boa ideia descrever as unidades das variáveis aqui.
\describe{
  \item{Nome da variável no dataset (seu tipo)}{Sua descrição}
  \item{municipio (factor)}{municipio do Estado do Paraná}
  \item{macro (factor)}{Umas das 4 regiões do Estado do Paraná (OESTE,LESTE,)}
  ...
}
}
\source{
fornece detalhes de onde você obteve os dados, geralmente um \url{http://www.diamondse.info/}
}
\usage{
dados_sinasc
}
\description{
Dataset usado na Aba Análise descritiva do Dashboard
}
\keyword{datasets}
