% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_sinasc_intervencao.R
\docType{data}
\name{dados_sinasc_intervencao}
\alias{dados_sinasc_intervencao}
\title{Dados do SINASC (Sistema de Informações sobre Nascidos Vivos)}
\format{
forneça uma visão geral do conjunto de dados. Para `data.frames`,
 você deve incluir uma lista de definições que descreva cada variável.
 Geralmente é uma boa ideia descrever as unidades das variáveis aqui.
\describe{
  \item{Nome da variável no dataset (seu tipo)}{Sua descrição}
}
}
\source{
fornece detalhes de onde você obteve os dados, geralmente um \url{http://www.diamondse.info/}
}
\usage{
dados_sinasc_intervencao
}
\description{
Dataset usado na Aba Análise de intervenção do Dashboard
}
\keyword{datasets}
