% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analise_geografica_fct_mapas.R
\name{RSmapOrd}
\alias{RSmapOrd}
\title{RSmapOrd}
\usage{
RSmapOrd(varToPlot, legeName, mun, RS, legeLabels, plot.action = TRUE)
}
\arguments{
\item{varToPlot}{Categorical Information to be plotted}

\item{legeName}{Name for the legend}

\item{mun}{Municipality names}

\item{RS}{Center Health number (Character)}

\item{legeLabels}{Labels for the legend}

\item{plot.action}{True or False for plotting the graphic. The graph object is returned anyway.}
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
