% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analise_geografica_fct_mapas.R
\name{UFmapOrd}
\alias{UFmapOrd}
\title{UFmapOrd}
\usage{
UFmapOrd(varToPlot, legeName, mun, legeLabels, plot.action = TRUE)
}
\arguments{
\item{varToPlot}{Ordinal Information to be plotted}

\item{legeName}{Name for the legend}

\item{mun}{Municipality names}

\item{legeLabels}{Labels for the legend}

\item{plot.action}{True or False for plotting the graphic. The graph object is returned anyway.}
}
\value{
The return value, if any, from executing the function.
}
\description{
A fct function
}
